renamed zne_utils
